Mova 105 - Implement Explore Screen #12
- Add info bottom sheet

+++ shared/src/commonMain/kotlin/feature_detail/data/movie/remote/dto/MovieDetailDto.kt
@@ -25,7 +25,7 @@ data class MovieDetailDto(
     @SerialName("production_companies") val productionCompanies: List<ProductionCompanyDto>,
     @SerialName("production_countries") val productionCountries: List<ProductionCountry>,
     @SerialName("release_date") val releaseDate: String?,
-    val revenue: Int?,
+    val revenue: Double?,
     val runtime: Int?,
     @SerialName("spoken_languages") val spokenLanguages: List<SpokenLanguageDto>,
     val status: String?,
